fix(msvc): fixed compilation with msvc

+++ Holovibes/cuda_includes/tools_conversion.cuh
@@ -20,7 +20,7 @@ void complex_to_modulus(float* output,
                         const cuComplex* input,
                         const ushort pmin,
                         const ushort pmax,
-                        const uint size,
+                        const size_t size,
                         const cudaStream_t stream);
 
 /*! \brief Compute the squared modulus of complex image(s).
@@ -34,7 +34,7 @@ void complex_to_squared_modulus(float* output,
                                 const cuComplex* input,
                                 const ushort pmin,
                                 const ushort pmax,
-                                const uint size,
+                                const size_t size,
                                 const cudaStream_t stream);
 
 /*! \brief Compute argument (angle) of complex image(s).
@@ -48,7 +48,7 @@ void complex_to_argument(float* output,
                          const cuComplex* input,
                          const ushort pmin,
                          const ushort pmax,
-                         const uint size,
+                         const size_t size,
                          const cudaStream_t stream);
 
 /*! \brief Copy the rescaled angle value of each element of the input.
@@ -62,7 +62,7 @@ void complex_to_argument(float* output,
  * \param size The number of elements to process.
  * \param stream The CUDA stream on which to launch the operation.
  */
-void rescale_float(const float* input, float* output, const uint size, const cudaStream_t stream);
+void rescale_float(const float* input, float* output, const size_t size, const cudaStream_t stream);
 
 /*! \brief Convert from big endian to little endian.
  * \param input The input data in big endian.
@@ -71,7 +71,7 @@ void rescale_float(const float* input, float* output, const uint size, const cud
  * \param stream The CUDA stream on which to launch the operation.
  */
 void endianness_conversion(
-    const ushort* input, ushort* output, const uint batch_size, const uint frame_res, const cudaStream_t stream);
+    const ushort* input, ushort* output, const uint batch_size, const size_t frame_res, const cudaStream_t stream);
 
 /*! \brief Convert data from float to unsigned short (16-bit).
  *
@@ -86,9 +86,9 @@ void endianness_conversion(
  * \param stream The CUDA stream on which to launch the operation.
  */
 void float_to_ushort(
-    const float* const input, ushort* const output, const uint size, const cudaStream_t stream, const uint shift = 0);
+    const float* const input, ushort* const output, const size_t size, const cudaStream_t stream, const uint shift = 0);
 
-void ushort_to_uchar(const ushort* input, uchar* output, const uint size, const cudaStream_t stream);
+void ushort_to_uchar(const ushort* input, uchar* output, const size_t size, const cudaStream_t stream);
 
 /*! \brief Converts and tranfers data from input_queue to gpu_input_buffer
  *
@@ -104,7 +104,7 @@ void ushort_to_uchar(const ushort* input, uchar* output, const uint size, const 
  */
 void input_queue_to_input_buffer(void* const output,
                                  const void* const input,
-                                 const uint frame_res,
+                                 const size_t frame_res,
                                  const int batch_size,
                                  const uint depth,
                                  const cudaStream_t stream);
@@ -144,11 +144,16 @@ __global__ void kernel_accumulate_images(const float* input,
                                          const size_t nb_elmt,
                                          const size_t nb_pixel);
 
-void normalize_complex(cuComplex* image, const uint size, const cudaStream_t stream);
+void normalize_complex(cuComplex* image, const size_t size, const cudaStream_t stream);
 
-void rescale_float_unwrap2d(float* input, float* output, float* cpu_buffer, uint frame_res, const cudaStream_t stream);
+void rescale_float_unwrap2d(
+    float* input, float* output, float* cpu_buffer, size_t frame_res, const cudaStream_t stream);
 
-void convert_frame_for_display(
-    const void* input, void* output, const uint size, const uint depth, const ushort shift, const cudaStream_t stream);
+void convert_frame_for_display(const void* input,
+                               void* output,
+                               const size_t size,
+                               const uint depth,
+                               const ushort shift,
+                               const cudaStream_t stream);
 
 void float_to_complex(cuComplex* output, const float* input, size_t size, const cudaStream_t stream);
