+++ Holovibes/cuda_sources/composite.cu
@@ -157,7 +157,7 @@ __global__ static void kernel_precompute_colors(float* colors,
 
 void rgb(cuComplex* input,
          float* output,
-         const uint frame_res,
+         const size_t frame_res,
          bool normalize,
          const ushort red,
          const ushort blue,
@@ -194,7 +194,7 @@ void rgb(cuComplex* input,
 }
 
 void postcolor_normalize(float* output,
-                         const uint frame_res,
+                         const size_t frame_res,
                          const uint real_line_size,
                          holovibes::units::RectFd selection,
                          const float weight_r,
