+++ Holovibes/sources/compute/fourier_transform.cc
@@ -216,7 +216,7 @@ void FourierTransform::insert_pca()
             cuComplex* V = nullptr;
 
             // cov = H' * H
-            cov_matrix(H, fd_.get_frame_res(), cd_.time_transformation_size, cov);
+            cov_matrix(H, static_cast<int>(fd_.get_frame_res()), cd_.time_transformation_size, cov);
 
             // Find eigen values and eigen vectors of cov
             // pca_eigen_values will contain sorted eigen values
@@ -232,7 +232,7 @@ void FourierTransform::insert_pca()
             // gpu_p_acc_buffer = H * V
             matrix_multiply(H,
                             V,
-                            fd_.get_frame_res(),
+                            static_cast<int>(fd_.get_frame_res()),
                             cd_.time_transformation_size,
                             cd_.time_transformation_size,
                             time_transformation_env_.gpu_p_acc_buffer);
@@ -255,7 +255,7 @@ void FourierTransform::insert_ssa_stft()
             cuComplex* V = nullptr;
 
             // cov = H' * H
-            cov_matrix(H, fd_.get_frame_res(), cd_.time_transformation_size, cov);
+            cov_matrix(H, static_cast<int>(fd_.get_frame_res()), cd_.time_transformation_size, cov);
 
             // pca_eigen_values = sorted eigen values of cov
             // cov and V = eigen vectors of cov
@@ -290,7 +290,7 @@ void FourierTransform::insert_ssa_stft()
             // H = H * tmp
             matrix_multiply(H,
                             tmp_matrix,
-                            fd_.get_frame_res(),
+                            static_cast<int>(fd_.get_frame_res()),
                             cd_.time_transformation_size,
                             cd_.time_transformation_size,
                             time_transformation_env_.gpu_p_acc_buffer);
@@ -306,7 +306,7 @@ void FourierTransform::insert_store_p_frame()
     fn_compute_vect_.conditional_push_back(
         [=]()
         {
-            const int frame_res = fd_.get_frame_res();
+            const int frame_res = static_cast<int>(fd_.get_frame_res());
 
             /* Copies with DeviceToDevice (which is the case here) are asynchronous
              * with respect to the host but never overlap with kernel execution*/
