+++ Holovibes/unit_tests/test_batch_input_queue.cc
@@ -23,13 +23,14 @@ static void ASSERT_QUEUE_ELT_EQ(holovibes::BatchInputQueue& q, size_t pos, std::
 
 static char* dequeue_helper(holovibes::BatchInputQueue& q, uint batch_size)
 {
-    const uint frame_size = q.get_fd().get_frame_size();
+    const size_t frame_size = q.get_fd().get_frame_size();
     static const holovibes::BatchInputQueue::dequeue_func_t lambda = [](const void* const src,
                                                                         void* const dest,
                                                                         const uint batch_size,
-                                                                        const uint frame_res,
+                                                                        const size_t frame_res,
                                                                         const uint depth,
-                                                                        const cudaStream_t stream) {
+                                                                        const cudaStream_t stream)
+    {
         const size_t size = static_cast<size_t>(batch_size) * frame_res * depth;
         cudaSafeCall(cudaMemcpyAsync(dest, src, size, cudaMemcpyDeviceToHost, stream));
     };
@@ -61,7 +62,7 @@ TEST(BatchInputQueueTest, SimpleEnqueueOfThreeElements)
     constexpr uint batch_size = 1;
     constexpr camera::FrameDescriptor fd = {2, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "a\0b\0c\0d\0e\0";
 
@@ -82,7 +83,7 @@ TEST(BatchInputQueueTest, SimpleEnqueueAndDequeueOfThreeElements)
     constexpr uint batch_size = 1;
     constexpr camera::FrameDescriptor fd = {2, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "a\0b\0c\0d\0e\0";
 
@@ -110,7 +111,7 @@ TEST(BatchInputQueueTest, SimpleOverwriteElements)
     constexpr uint batch_size = 1;
     constexpr camera::FrameDescriptor fd = {2, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "a\0b\0c\0d\0e\0";
 
@@ -141,7 +142,7 @@ TEST(BatchInputQueueTest, SimpleOverwriteMoreElements)
     constexpr uint batch_size = 2;
     constexpr camera::FrameDescriptor fd = {4, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "abc\0ABC\0def\0DEF\0ghi\0GHI\0";
 
@@ -194,7 +195,7 @@ TEST(BatchInputQueueTest, SimpleResizeSame)
     constexpr uint batch_size = 2;
     constexpr camera::FrameDescriptor fd = {5, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "ilan\0nico\0anto\0kaci\0theo\0";
 
@@ -225,7 +226,7 @@ TEST(BatchInputQueueTest, SimpleResizeGreater)
     constexpr uint batch_size = 2;
     constexpr camera::FrameDescriptor fd = {5, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "ilan\0nico\0anto\0kaci\0theo\0";
 
@@ -258,7 +259,7 @@ TEST(BatchInputQueueTest, SimpleResizeLower)
     constexpr uint batch_size = 2;
     constexpr camera::FrameDescriptor fd = {5, 1, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_size = queue.get_fd().get_frame_size();
+    size_t frame_size = queue.get_fd().get_frame_size();
 
     const char* data = "ilan\0nico\0anto\0kaci\0theo\0";
 
@@ -336,7 +337,7 @@ void consumer_gpu(holovibes::BatchInputQueue& queue,
 }
 
 template <typename T>
-void producer(holovibes::BatchInputQueue& queue, const uint nb_actions, const uint frame_res)
+void producer(holovibes::BatchInputQueue& queue, const uint nb_actions, const size_t frame_res)
 {
     const T* frame = new T[frame_res];
 
@@ -350,7 +351,7 @@ void producer(holovibes::BatchInputQueue& queue, const uint nb_actions, const ui
 
 void producer_gpu(holovibes::BatchInputQueue& queue,
                   const uint nb_actions,
-                  const uint frame_res,
+                  const size_t frame_res,
                   const float* const d_buff)
 {
 
@@ -372,7 +373,7 @@ TEST(BatchInputQueueTest, ProducerConsumerSituationNoDeadlock)
         constexpr uint max_batch_size = total_nb_frames;
         constexpr camera::FrameDescriptor fd = {2, 2, sizeof(float), camera::Endianness::LittleEndian};
         holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-        uint frame_res = queue.get_fd().get_frame_res();
+        size_t frame_res = queue.get_fd().get_frame_res();
 
         // Consumer will do less actions. It is maximum in case of batch size ==
         // 1
@@ -404,13 +405,14 @@ TEST(BatchInputQueueTest, ProducerConsumerSituationNoDeadlockSmallSize)
     constexpr camera::FrameDescriptor fd = {1, 1, sizeof(float), camera::Endianness::LittleEndian};
     constexpr uint total_nb_frames = 2;
     constexpr uint batch_size = 1;
-    const uint frame_size = fd.get_frame_size();
+    const size_t frame_size = fd.get_frame_size();
     static const holovibes::BatchInputQueue::dequeue_func_t dequeue_func = [](const void* const src,
                                                                               void* const dest,
                                                                               const uint batch_size,
-                                                                              const uint frame_res,
+                                                                              const size_t frame_res,
                                                                               const uint depth,
-                                                                              const cudaStream_t stream) {
+                                                                              const cudaStream_t stream)
+    {
         const size_t size = static_cast<size_t>(batch_size) * frame_res * depth;
         cudaSafeCall(cudaMemcpyAsync(dest, src, size, cudaMemcpyDeviceToDevice, stream));
     };
@@ -424,7 +426,7 @@ TEST(BatchInputQueueTest, ProducerConsumerSituationNoDeadlockSmallSize)
     for (uint i = 0; i < nb_tests; i++)
     {
         holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-        uint frame_res = queue.get_fd().get_frame_res();
+        size_t frame_res = queue.get_fd().get_frame_res();
 
         // Consumer will do less actions. It is maximum in case of batch size ==
         // 1
@@ -459,7 +461,7 @@ TEST(BatchInputQueueTest, FullProducerConsumerSituationFloat)
     constexpr uint max_batch_size = total_nb_frames;
     constexpr camera::FrameDescriptor fd = {2, 2, sizeof(float), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_res = queue.get_fd().get_frame_res();
+    size_t frame_res = queue.get_fd().get_frame_res();
 
     // Consumer will do less actions. It is maximum in case of batch size ==
     // 1
@@ -496,7 +498,7 @@ TEST(BatchInputQueueTest, FullProducerConsumerSituationChar)
     constexpr uint max_batch_size = total_nb_frames;
     constexpr camera::FrameDescriptor fd = {2, 2, sizeof(char), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_res = queue.get_fd().get_frame_res();
+    size_t frame_res = queue.get_fd().get_frame_res();
 
     // Consumer will do less actions. It is maximum in case of batch size ==
     // 1
@@ -533,7 +535,7 @@ TEST(BatchInputQueueTest, FullProducerConsumerSituationShort)
     constexpr uint max_batch_size = total_nb_frames;
     constexpr camera::FrameDescriptor fd = {2, 2, sizeof(short), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_res = queue.get_fd().get_frame_res();
+    size_t frame_res = queue.get_fd().get_frame_res();
 
     // Consumer will do less actions. It is maximum in case of batch size ==
     // 1
@@ -570,7 +572,7 @@ TEST(BatchInputQueueTest, PartialProducerConsumerSituationShort)
     constexpr uint max_batch_size = total_nb_frames;
     constexpr camera::FrameDescriptor fd = {2, 2, sizeof(short), camera::Endianness::LittleEndian};
     holovibes::BatchInputQueue queue(total_nb_frames, batch_size, fd);
-    uint frame_res = queue.get_fd().get_frame_res();
+    size_t frame_res = queue.get_fd().get_frame_res();
 
     // Consumer will do less actions. It is maximum in case of batch size ==
     // 1
